git testes added

+++ Libft/ft_bzero.c
@@ -6,15 +6,16 @@
 /*   By: agorski <agorski@student.42warsaw.pl>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/01 18:16:02 by agorski           #+#    #+#             */
-/*   Updated: 2024/03/14 14:29:46 by agorski          ###   ########.fr       */
+/*   Updated: 2024/03/14 16:54:41 by agorski          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
 	while (n--)
 	{
 		*(unsigned char *)s = '\0';
 	}
 }
-//test
