+++ Libft/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: agorski <agorski@student.42warsaw.pl>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/02 20:50:12 by agorski           #+#    #+#             */
-/*   Updated: 2024/03/13 17:46:56 by agorski          ###   ########.fr       */
+/*   Updated: 2024/03/14 17:43:12 by agorski          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	char		*dest_temp;
 	const char	*src_temp;
