+++ Libft/libft.h
@@ -6,23 +6,24 @@
 /*   By: agorski <agorski@student.42warsaw.pl>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/06 18:24:01 by agorski           #+#    #+#             */
-/*   Updated: 2024/03/11 19:08:49 by agorski          ###   ########.fr       */
+/*   Updated: 2024/03/14 18:03:17 by agorski          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
-#define LIBFT_H
+# define LIBFT_H
+# include <stdlib.h>
 
-int	ft_isalpha(int c);
-int	ft_isdigit(int c);
-int	ft_isalnum(int c);
-int	ft_isascii(int c);
-int	ft_isprint(int c);
-int	ft_strlen(const char *str);
-void	*ft_memset(void *s, int c, unsigned int n);
-void	ft_bzero(void *s, unsigned int n);
-void	*ft_memcpy(void *dest, const void *src, unsigned int n);
-void	*ft_memmove(void *dest, const void *src, unsigned int n);
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size);
+int				ft_isalpha(int c);
+int				ft_isdigit(int c);
+int				ft_isalnum(int c);
+int				ft_isascii(int c);
+int				ft_isprint(int c);
+size_t				ft_strlen(const char *str);
+void			*ft_memset(void *s, int c, size_t n);
+void			ft_bzero(void *s, size_t n);
+void			*ft_memcpy(void *dest, const void *src, size_t n);
+void			*ft_memmove(void *dest, const void *src, size_t n);
+size_t	ft_strlcpy(char *dest, char *src, size_t size);
 
 #endif
