+++ Libft_test/test_ft_strlcat.c
@@ -2,7 +2,7 @@
 #include <string.h>
 
 // Function prototype
-unsigned int ft_strlcat(char *dst, const char *src, unsigned int size);
+size_t ft_strlcat(char *dst, const char *src, size_t size);
 
 int main(void)
 {
